requirements.txt
python-telegram-bot==21.5
telethon==1.36.0
python-dotenv==1.0.0
httpx==0.27.0
requests=2.33.1